use more explicit labels and rename folders
